[hir2mpl]Internal commit msg:     [DTS2022012712052] bugfix deal with 0xdeadbeef in lower and upper bound [mapleall]Internal commit msg:     [mapleall] pclint waring clean14     [mplcg] check cases for reconstructed patterns based-on ssa     [DTS2022012712052] bugfix deal with 0xdeadbeef in lower and upper bound

+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_optimize_common.cpp
@@ -50,8 +50,7 @@ void AArch64InsnVisitor::ModifyJumpTarget(maple::LabelIdx targetLabel, BB &bb) {
 }
 
 void AArch64InsnVisitor::ModifyJumpTarget(BB &newTarget, BB &bb) {
-  ModifyJumpTarget(newTarget.GetLastInsn()->GetOperand(
-      static_cast<int32>(newTarget.GetLastInsn()->GetJumpTargetIdx())), bb);
+  ModifyJumpTarget(newTarget.GetLastInsn()->GetOperand(newTarget.GetLastInsn()->GetJumpTargetIdx()), bb);
 }
 
 Insn *AArch64InsnVisitor::CloneInsn(Insn &originalInsn) {
@@ -80,7 +79,7 @@ Insn *AArch64InsnVisitor::CloneInsn(Insn &originalInsn) {
  * because a register instead of label. So we don't take it as a branching instruction.
  */
 LabelIdx AArch64InsnVisitor::GetJumpLabel(const Insn &insn) const {
-  int operandIdx = static_cast<int>(insn.GetJumpTargetIdx());
+  uint32 operandIdx = insn.GetJumpTargetIdx();
   if (insn.GetOperand(operandIdx).IsLabelOpnd()) {
     return static_cast<LabelOperand&>(insn.GetOperand(operandIdx)).GetLabelIndex();
   }
