+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_ra_opt.cpp
@@ -39,7 +39,7 @@ bool RaX0Opt::PropagateRenameReg(Insn *nInsn, const X0OptInfo &optVal) {
   const AArch64MD *md = &AArch64CG::kMd[static_cast<AArch64Insn*> (nInsn)->GetMachineOpcode()];
   int32 lastOpndId = nInsn->GetOperandSize() - 1;
   for (int32_t i = lastOpndId; i >= 0; i--) {
-    Operand &opnd = nInsn->GetOperand(i);
+    Operand &opnd = nInsn->GetOperand(static_cast<uint32>(i));
 
     if (opnd.IsList()) {
       /* call parameters */
